reorganize Legend Query code
