+++ packages/legend-query/src/stores/__tests__/QueryBuilder_ExplorerState.test.ts
@@ -41,7 +41,7 @@ import {
 } from '../../stores/QueryBuilderExplorerState.js';
 import { LegendQueryPluginManager } from '../../application/LegendQueryPluginManager.js';
 import { QueryBuilder_GraphPreset } from '../../models/QueryBuilder_GraphPreset.js';
-import { TEST__provideMockedLegendQueryStore } from '../../components/QueryComponentTestUtils.js';
+import { TEST__provideMockedLegendQueryStore } from '../../components/LegendQueryComponentTestUtils.js';
 import {
   TEST_DATA__Auto_M2M,
   TEST_DATA__Relational_Inline,
