+++ packages/legend-query/style/_query-builder.scss
@@ -277,9 +277,10 @@
   }
 
   &__analytics {
-    font-size: 1.2rem;
-    color: var(--color-light-grey-300);
+    font-size: 1.1rem;
+    color: var(--color-dark-grey-400);
     margin-left: 0.5rem;
+    font-family: 'Roboto Mono', monospace;
   }
 
   &__limit {
