+++ packages/legend-studio/style/components/editor/_service-editor.scss
@@ -216,7 +216,6 @@
     @include flexVCenter;
 
     height: 100%;
-    width: 12rem;
     margin: 0 0.3rem;
 
     &__label {
